+++ objects_mpi/Makefile
@@ -1,5 +1,4 @@
 # -*- Makefile for unit test -*-
-SHELL=/bin/sh
 CC=mpic++
 CFLAGS=-Wall -g -O3 -ffast-math -fomit-frame-pointer -DUSE_MPI #-DTIMING
 LDLIBS=-lm
@@ -16,16 +15,12 @@ HEADS=$(wildcard $(INCDIR)*.h)
 default: ../ljmd-mpi.x
 
 # linker rule
-../ljmd-mpi.x: $(OBJ) $(SRCDIR)/cell.o
+../ljmd-mpi.x: $(OBJ)
 	$(CC) -o $@ $(CFLAGS) $^ $(LDLIBS)
 
 # compile true source files
 $(OBJDIR)/%.o: $(SRCDIR)/%.c $(HEADS)
 	$(CC) -c $(CFLAGS) $< -o $@
 
-# special for cell
-$(OBJDIR)/cell.o: $(SRCDIR)/%.cc $(HEADS)
-	$(CC) -c $(CFLAGS) $< -o $@
-
 clean:
-	rm -f ../ljmd-mpi.x *.o
+	rm -f ../ljmd-mpi.x $(OBJDIR)/*.o
