+++ objects_serial/Makefile
@@ -1,5 +1,4 @@
 # -*- Makefile -*-
-SHELL=/bin/sh
 CC=g++
 CFLAGS=-Wall -g -O3 -ffast-math -fomit-frame-pointer #-DTIMING #-pg
 LDLIBS=-lm
@@ -9,23 +8,18 @@ SRCDIR=../src
 OBJDIR=./
 SOURCES= $(shell find $(SRCDIR) -type f -name *.c)
 OBJ= $(patsubst $(SRCDIR)/%,./%,$(SOURCES:.c=.o))
-#FOBJ= $(filter-out $(OBJDIR)/ljmd.o,$(OBJ))
 
 default: ../ljmd-serial.x
 
 clean:
-	rm -f *.mod *.o ../ljmd-serial.x
+	rm -f *.o ../ljmd-serial.x
 
 # compilation pattern rule for objects
 $(OBJDIR)/%.o: $(SRCDIR)/%.c
 	$(CC) -c $(CFLAGS) $< -o $@
 
-# special for cell
-$(OBJDIR)/cell.o: $(SRCDIR)/cell.cc $(SRCDIR)/cell.h
-	$(CC) -c $(CFLAGS) -I$(SRCDIR) -I. $< -o $@
-
 # linker rule
-../ljmd-serial.x: $(OBJ) cell.o
+../ljmd-serial.x: $(OBJ)
 	$(CC) -o $@ $(CFLAGS) $^ $(LDLIBS)
