+++ opt/gprof.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
-cd ..
-make
+make serial
 cd examples
 rm gmon.out
 ../ljmd-serial.x < argon_108.inp
